Update TMs + HMs
300 TMS + 8 HMs
Expanded bag pocket slots

+++ src/rom_header_gf.c
@@ -80,10 +80,10 @@ struct GFRomHeader
     u32 gcnLinkFlagsOffset;
     u32 gameClearFlag;
     u32 ribbonFlag;
-    u8 bagCountItems;
+    u16 bagCountItems;
     u8 bagCountKeyItems;
     u8 bagCountPokeballs;
-    u8 bagCountTMHMs;
+    u16 bagCountTMHMs;
     u8 bagCountBerries;
     u8 pcItemsCount;
     u32 pcItemsOffset;
